backend/requirements.txt
fastapi==0.110.0
uvicorn==0.27.1
motor==3.3.2
pydantic==2.6.3
python-dotenv==1.0.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.9
email-validator==2.1.1